requirements.txt
sentence-transformers==2.2.2
scikit-learn==1.2.2
nltk==3.8.1
python-docx==0.8.11
numpy==1.24.3
pandas==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0